requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.31.0
python-dotenv==1.0.0
sqlmodel==0.0.16
python-multipart==0.0.7
pdfplumber==0.10.4
pandas==2.2.2